Fail if we're not running from a clean working directory
"olx new-run" previously did not care whether there were uncommitted
changes, or untracked files, in the working directory. This was
almost always a bad idea for running with -b, because it meant
that changes may have been included in a run branch when they really
belonged into the parent (usually 'master') branch, or that files
were added to a run that didn't belong there at all.

Now, just bail if the working directory is dirty.